refactor des lstc compiling
